refactor: extract shared useDismiss hook for popovers and menus
The same close-on-outside-press-or-Escape block was re-implemented in
five places (Select, the sort dropdown, the sidebar identity menu, and
the coordinate-positioned note/session and meeting row context menus).
One hook now owns it: containment-checked when a wrapper ref is passed,
close-on-any-press when not, with the pointer event configurable
because the coordinate menus listen on click rather than mousedown.
Zero behavior change per site; documented as the canonical dismiss in
docs/design/components.md.
